requirements.txt
retry==0.9.2
pyyaml==6.0
gitpython==3.1.27
algorithmia==1.16.1
datarobot==2.27.1